pyproject.toml
[tool.poetry]
name = "Crypto Prices"
version = "1.6.0"
description = "Get current prices for cryptocurrencies"
documentation = "https://github.com/fedecalendino/alfred-crypto-prices/blob/main/README.md"
homepage = "https://github.com/fedecalendino/alfred-crypto-prices"
license = "MIT"

authors = [
    "Fede Calendino <fede@calendino.com>"
]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
alfred-pyflow = "latest"
pycoingecko = "^3.0.0"

[tool.poetry.dev-dependencies]
black = "22.10.0"
pyinstaller = "5.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
